IVY-1631: when the ivy:retrieve Ant task retrieves the jars to a non-empty directory, but the retrieve task itself didn't retrieve any files, the created fileset was not empty but contained the existing files in that directory.
